Working prototype

+++ include/typedefs.h
@@ -1,7 +1,7 @@
 #ifndef __TYPEDEFS_H__
 #define __TYPEDEFS_H__
 
-using meshInt = short unsigned int;
+using meshInt   = unsigned int;
 using coordType = double;
 
 #endif // __TYPEDEFS_H__
